break up our bootstrap and our initialization
